Reject some invalid cfg usage
single-core related cfgs (disable_fiq and s_mode) were accidentally
allowed in 0.3.16-0.3.18 and were no-op. However, this is considered bug
fix because single-core cfg explicitly states that using cfg on
unsupported platforms will result in an error.

Given that no one on GitHub appears to be using these, I will include
this in the patch release, but please open an issue if you encounter any
problems.