tests: Implement mechanism to detect API changes by OS
Run-time feature detection on aarch64 usually requires the use of
platform APIs, and we define our own FFI bindings to those APIs.

We use only one function and a few constants per platform, but
technically OS can change these APIs so it is preferable to be able
to detect them.