Remove uses of unreachable_unchecked
Eliminates branching within the called function by moving the branch to
the caller.

This removes the need for unreachable_unchecked macro and the
assert_swap_ordering function required to use it correctly, and also
slightly improves performance (1-2% on aarch64).

This is also consistent with the way GCC/LLVM's outline-atomics works.