Re-implement on top of v1.x
The latest version of portable-atomic is 1.x. This patch makes
portable-atomic 0.3 is built top of portable-atomic 1.x to make bug
fixes and improvements such as support for new targets in 1.x available
to the ecosystem that depends on older portable-atomic. portable-atomic
0.3 is still maintained passively, but upgrading to portable-atomic 1.x
is recommended. (There are no breaking changes from 0.3, except that a
deprecated no-op `outline-atomics` Cargo feature has been removed.)

This is intentionally not implemented as a re-export. Because it breaks
code that assumes 0.3 and 1.0 types are different types, e.g., the same
trait is implemented in both types.